feat: 상단 날짜표시 컴포넌트 분리
